refactor: Separate UpdatePasswordForm for improved content streaming
